@@ -2,20 +2,22 @@
 import React, {Component} from 'react';
 import {connect} from 'react-redux';
 
-import DebtCollection from './DebtCollection';
+import Button from '$components/button/Button';
 import Collapse from '$components/collapse/Collapse';
+import ConfirmationModal from '$components/modal/ConfirmationModal';
+import DebtCollection from './DebtCollection';
 import Divider from '$components/content/Divider';
 import InvoiceSimulator from '$components/invoice-simulator/InvoiceSimulator';
-import InvoicesTable from './InvoicesTable';
-import CreateAndCreditInvoiceComponent from './CreateAndCreditInvoiceComponent';
+import CreateAndCreditInvoice from './CreateAndCreditInvoice';
 import CreateCollectionLetter from './CreateCollectionLetter';
 import RentCalculator from '$components/rent-calculator/RentCalculator';
 import RightSubtitle from '$components/content/RightSubtitle';
+import InvoiceTableAndPanel from './InvoiceTableAndPanel';
 import {receiveInvoiceToCredit, receiveIsCreateInvoicePanelOpen, receiveIsCreditInvoicePanelOpen} from '$src/invoices/actions';
-import {receiveCollapseStates} from '$src/leases/actions';
+import {receiveCollapseStates, startInvoicing, stopInvoicing} from '$src/leases/actions';
 import {ViewModes} from '$src/enums';
 import {getInvoiceToCredit} from '$src/invoices/selectors';
-import {getCollapseStateByKey, getCurrentLease} from '$src/leases/selectors';
+import {getCollapseStateByKey, getCurrentLease, getIsEditMode} from '$src/leases/selectors';
 
 import type {Lease} from '$src/leases/types';
 
@@ -23,6 +25,7 @@ type Props = {
   currentLease: Lease,
   invoicesCollapseState: boolean,
   invoiceToCredit: ?string,
+  isEditMode: boolean,
   isInvoicingEnabled: boolean,
   previewInvoicesCollapseState: boolean,
   receiveCollapseStates: Function,
@@ -30,16 +33,42 @@ type Props = {
   receiveIsCreditInvoicePanelOpen: Function,
   receiveInvoiceToCredit: Function,
   rentCalculatorCollapseState: boolean,
+  startInvoicing: Function,
+  stopInvoicing: Function,
 }
 
-class Invoices extends Component<Props> {
+type State = {
+  isStartInvoicingModalOpen: boolean,
+  isStopInvoicingModalOpen: boolean,
+}
+
+class Invoices extends Component<Props, State> {
   creditPanel: any
 
-  componentWillMount = () => {
+  state = {
+    isStartInvoicingModalOpen: false,
+    isStopInvoicingModalOpen: false,
+  }
+
+  componentDidMount = () => {
     const {receiveInvoiceToCredit, receiveIsCreateInvoicePanelOpen, receiveIsCreditInvoicePanelOpen} = this.props;
     receiveIsCreateInvoicePanelOpen(false);
     receiveIsCreditInvoicePanelOpen(false);
     receiveInvoiceToCredit(null);
+  }
+
+  showModal = (modalName: string) => {
+    const modalVisibilityKey = `is${modalName}ModalOpen`;
+    this.setState({
+      [modalVisibilityKey]: true,
+    });
+  }
+
+  hideModal = (modalName: string) => {
+    const modalVisibilityKey = `is${modalName}ModalOpen`;
+    this.setState({
+      [modalVisibilityKey]: false,
+    });
   }
 
   handleInvoicesCollapseToggle = (val: boolean) => {
@@ -83,38 +112,114 @@ class Invoices extends Component<Props> {
     receiveInvoiceToCredit(val);
   };
 
+  handleStartInvoicingButtonClick = () => {
+    this.showModal('StartInvoicing');
+  }
+
+  handleStartInvoicingCancelButtonClick = () => {
+    this.hideModal('StartInvoicing');
+  }
+
+  handleStopInvoicingButtonClick = () => {
+    this.showModal('StopInvoicing');
+  }
+
+  handleStopInvoicingCancelButtonClick = () => {
+    this.hideModal('StopInvoicing');
+  }
+
+  startInvoicing = () => {
+    const {currentLease, startInvoicing} = this.props;
+
+    this.hideModal('StartInvoicing');
+    startInvoicing(currentLease.id);
+  }
+
+  stopInvoicing = () => {
+    const {currentLease, stopInvoicing} = this.props;
+
+    this.hideModal('StopInvoicing');
+    stopInvoicing(currentLease.id);
+  }
+
   render() {
     const {
       invoicesCollapseState,
       invoiceToCredit,
+      isEditMode,
       isInvoicingEnabled,
       previewInvoicesCollapseState,
       rentCalculatorCollapseState,
     } = this.props;
+    const {
+      isStartInvoicingModalOpen,
+      isStopInvoicingModalOpen,
+    } = this.state;
 
     return (
       <div>
-        <h2>Laskutus</h2>
-        <RightSubtitle
-          className='invoicing-status'
-          text={isInvoicingEnabled
-            ? <p className="success">Laskutus käynnissä</p>
-            : <p className="alert">Laskutus ei käynnissä</p>
-          }
+        <ConfirmationModal
+          confirmButtonLabel='Käynnistä laskutus'
+          isOpen={isStartInvoicingModalOpen}
+          label='Haluatko varmasti käynnistää laskutuksen?'
+          onCancel={this.handleStartInvoicingCancelButtonClick}
+          onClose={this.handleStartInvoicingCancelButtonClick}
+          onSave={this.startInvoicing}
+          title='Käynnistä laskutus'
         />
+        <ConfirmationModal
+          confirmButtonLabel='Keskeytä laskutus'
+          isOpen={isStopInvoicingModalOpen}
+          label='Haluatko varmasti keskeyttää laskutuksen?'
+          onCancel={this.handleStopInvoicingCancelButtonClick}
+          onClose={this.handleStopInvoicingCancelButtonClick}
+          onSave={this.stopInvoicing}
+          title='Keskeytä laskutus'
+        />
+
+        <h2>Laskutus</h2>
+        {isEditMode
+          ? <RightSubtitle
+            buttonComponent={isInvoicingEnabled
+              ? <Button
+                className='button-red'
+                onClick={this.handleStopInvoicingButtonClick}
+                text='Keskeytä laskutus'
+              />
+              : <Button
+                className='button-green'
+                onClick={this.handleStartInvoicingButtonClick}
+                text='Käynnistä laskutus'
+              />
+            }
+            className='invoicing-status'
+            text={isInvoicingEnabled
+              ? <p className="success">Laskutus käynnissä</p>
+              : <p className="alert">Laskutus ei käynnissä</p>
+            }
+          />
+
+          : <RightSubtitle
+            className='invoicing-status'
+            text={isInvoicingEnabled
+              ? <p className="success">Laskutus käynnissä</p>
+              : <p className="alert">Laskutus ei käynnissä</p>
+            }
+          />
+        }
+
         <Divider />
         <Collapse
           defaultOpen={invoicesCollapseState !== undefined ? invoicesCollapseState : true}
           headerTitle={<h3 className='collapse__header-title'>Laskut</h3>}
           onToggle={this.handleInvoicesCollapseToggle}
         >
-          <InvoicesTable
+          <InvoiceTableAndPanel
             invoiceToCredit={invoiceToCredit}
             onInvoiceToCreditChange={this.handleInvoiceToCreditChange}
           />
-
-          <CreateAndCreditInvoiceComponent
-            enableCreateInvoice={false}
+          <CreateAndCreditInvoice
+            enableCreateInvoice={isEditMode}
             enableCreditInvoice={true}
             invoiceToCredit={invoiceToCredit}
           />
@@ -150,6 +255,7 @@ export default connect(
       currentLease: currentLease,
       invoicesCollapseState: getCollapseStateByKey(state, `${ViewModes.READONLY}.invoices.invoices`),
       invoiceToCredit: getInvoiceToCredit(state),
+      isEditMode: getIsEditMode(state),
       isInvoicingEnabled: currentLease ? currentLease.is_invoicing_enabled : null,
       previewInvoicesCollapseState: getCollapseStateByKey(state, `${ViewModes.READONLY}.invoices.preview_invoices`),
       rentCalculatorCollapseState: getCollapseStateByKey(state, `${ViewModes.READONLY}.invoices.rent_calculator`),
@@ -160,5 +266,7 @@ export default connect(
     receiveInvoiceToCredit,
     receiveIsCreateInvoicePanelOpen,
     receiveIsCreditInvoicePanelOpen,
+    startInvoicing,
+    stopInvoicing,
   }
 )(Invoices);
