@@ -2,13 +2,13 @@
 import React from 'react';
 
 export const SortIconBoth = () =>
-  <svg className='table__sortable_sort-icon' focusable='false' xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24"><path d="M12 0l8 10h-16l8-10zm8 14h-16l8 10 8-10z"/></svg>;
+  <svg className='sortable-table-sort-icon' focusable='false' xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24"><path d="M12 0l8 10h-16l8-10zm8 14h-16l8 10 8-10z"/></svg>;
 
 export const SortIconAsc = () =>
-  <svg className='table__sortable_sort-icon' focusable='false' xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24"><path d="M12 0l-8 10h16l-8-10zm3.839 16l-3.839 4.798-3.839-4.798h7.678zm4.161-2h-16l8 10 8-10z"/></svg>;
+  <svg className='sortable-table-sort-icon' focusable='false' xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24"><path d="M12 0l-8 10h16l-8-10zm3.839 16l-3.839 4.798-3.839-4.798h7.678zm4.161-2h-16l8 10 8-10z"/></svg>;
 
 export const SortIconDesc = () =>
-  <svg className='table__sortable_sort-icon' focusable='false' xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24"><path d="M12 3.202l3.839 4.798h-7.678l3.839-4.798zm0-3.202l-8 10h16l-8-10zm8 14h-16l8 10 8-10z"/></svg>;
+  <svg className='sortable-table-sort-icon' focusable='false' xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24"><path d="M12 3.202l3.839 4.798h-7.678l3.839-4.798zm0-3.202l-8 10h16l-8-10zm8 14h-16l8 10 8-10z"/></svg>;
 
 export const NextIcon = () =>
   <svg xmlns="http://www.w3.org/2000/svg"  focusable='false' width="14px" height="11px" viewBox="0 0 14 11">
