@@ -1,10 +1,20 @@
 // @flow
 /**
- * View modes enumerabke.
+ * View modes enumerable.
  *
  * @type {{EDIT: string, READONLY: string,}}
  */
 export const ViewModes = {
   EDIT: 'edit',
   READONLY: 'readonly',
+};
+
+/**
+ * Key code enumerable.
+ *
+ * @type {{ARROW_LEFT: string, ARROW_RIGHT: string}}
+ */
+export const KeyCodes = {
+  ARROW_LEFT: 37,
+  ARROW_RIGHT: 39,
 };
