@@ -2,7 +2,7 @@
 import React from 'react';
 import {connect} from 'react-redux';
 
-import Table from '$components/table/Table';
+import SortableTable from '$components/table/SortableTable';
 import {
   formatDate,
   formatNumber,
@@ -27,19 +27,17 @@ const IndexAdjustedRents = ({attributes, indexAdjustedRents}: Props) => {
     'rents.child.children.index_adjusted_rents.child.children.intended_use');
 
   return (
-    <Table
-      caption='Indeksitarkastetut vuokrat'
-      data={indexAdjustedRents}
-      dataKeys={[
-        {key: 'amount', label: 'Indeksitarkastettu vuokra', renderer: (val) => val ? `${formatNumber(val)} €` : '-', ascSortFunction: sortNumberByKeyAsc, descSortFunction: sortNumberByKeyDesc},
-        {key: 'intended_use', label: 'Käyttötarkoitus', renderer: (val) => getLabelOfOption(intendedUseOptions, val), ascSortFunction: (a, b, key) => sortByOptionsAsc(a, b, key, intendedUseOptions), descSortFunction: (a, b, key) => sortByOptionsDesc(a, b, key, intendedUseOptions)},
-        {key: 'start_date', label: 'Alkupvm', renderer: (val) => formatDate(val), defaultSorting: 'desc'},
-        {key: 'end_date', label: 'Loppupvm', renderer: (val) => formatDate(val)},
-        {key: 'factor', label: 'Laskentakerroin', renderer: (val) => formatNumber(val), ascSortFunction: sortNumberByKeyAsc, descSortFunction: sortNumberByKeyDesc},
+    <SortableTable
+      columns={[
+        {key: 'amount', text: 'Indeksitarkastettu vuokra', renderer: (val) => val ? `${formatNumber(val)} €` : '-', ascSortFunction: sortNumberByKeyAsc, descSortFunction: sortNumberByKeyDesc},
+        {key: 'intended_use', text: 'Käyttötarkoitus', renderer: (val) => getLabelOfOption(intendedUseOptions, val), ascSortFunction: (a, b, key) => sortByOptionsAsc(a, b, key, intendedUseOptions), descSortFunction: (a, b, key) => sortByOptionsDesc(a, b, key, intendedUseOptions)},
+        {key: 'start_date', text: 'Alkupvm', renderer: (val) => formatDate(val), defaultSorting: 'desc'},
+        {key: 'end_date', text: 'Loppupvm', renderer: (val) => formatDate(val)},
+        {key: 'factor', text: 'Laskentakerroin', renderer: (val) => formatNumber(val), ascSortFunction: sortNumberByKeyAsc, descSortFunction: sortNumberByKeyDesc},
       ]}
-      fixedHeader
-      sortable
-      tableFixedLayout
+      data={indexAdjustedRents}
+      fixedHeader={true}
+      sortable={true}
     />
   );
 };
