@@ -1,13 +1,13 @@
 // @flow
 import React from 'react';
 
+import SortableTable from '$components/table/SortableTable';
 import {
   formatDate,
   formatNumber,
   sortNumberByKeyAsc,
   sortNumberByKeyDesc,
 } from '$util/helpers';
-import Table from '$components/table/Table';
 
 type Props = {
   payableRents: Array<Object>,
@@ -15,19 +15,17 @@ type Props = {
 
 const PayableRents = ({payableRents}: Props) => {
   return (
-    <Table
-      caption='Perittävät vuokrat'
-      data={payableRents}
-      dataKeys={[
-        {key: 'amount', label: 'Indeksitarkastettu vuokra', renderer: (val) => val ? `${formatNumber(val)} €` : '-', ascSortFunction: sortNumberByKeyAsc, descSortFunction: sortNumberByKeyDesc},
-        {key: 'start_date', label: 'Alkupvm', renderer: (val) => formatDate(val), defaultSorting: 'desc'},
-        {key: 'end_date', label: 'Loppupvm', renderer: (val) => formatDate(val)},
-        {key: 'difference_percent', label: 'Nousu', renderer: (val) => val ? `${formatNumber(val)} %` : '-'},
-        {key: 'calendar_year_rent', label: 'Kalenterivuosivuokra', renderer: (val) => val ? `${formatNumber(val)} €` : '-'},
+    <SortableTable
+      columns={[
+        {key: 'amount', text: 'Indeksitarkastettu vuokra', renderer: (val) => val ? `${formatNumber(val)} €` : '-', ascSortFunction: sortNumberByKeyAsc, descSortFunction: sortNumberByKeyDesc},
+        {key: 'start_date', text: 'Alkupvm', renderer: (val) => formatDate(val), defaultSorting: 'desc'},
+        {key: 'end_date', text: 'Loppupvm', renderer: (val) => formatDate(val)},
+        {key: 'difference_percent', text: 'Nousu', renderer: (val) => val ? `${formatNumber(val)} %` : '-'},
+        {key: 'calendar_year_rent', text: 'Kalenterivuosivuokra', renderer: (val) => val ? `${formatNumber(val)} €` : '-'},
       ]}
-      fixedHeader
-      sortable
-      tableFixedLayout
+      data={payableRents}
+      fixedHeader={true}
+      sortable={true}
     />
   );
 };
